requirements.txt
PyQt6>=6.6.0
requests>=2.31.0
cryptography>=41.0.0
qrcode[pil]>=7.4.0
python-dotenv>=1.0.0
